Reduce string allocations on hot parse path.
When waiting for input we pass down strings for a possible error message
and the triggering location. In generated code these are always
literals.

With this patch we do not take them as owning strings, but instead as
views into existing strings to minimize allocations. In the case of
error messages the created low-level exception objects already had used
string_views, so this also aligns the APIs.

Closes #1589.